requirements.txt
django ==3.2
pillow ==8.1
django-crispy-forms==1.1
django-taggit ==1.4.0